+++ Sources/Views/ALKFriendPhotoCell.swift
@@ -38,8 +38,15 @@ class ALKFriendPhotoCell: ALKPhotoCell {
     
     override func setupStyle() {
         super.setupStyle()
-        
         nameLabel.setStyle(ALKMessageStyle.displayName)
+        if(ALKMessageStyle.receivedBubble.style == .edge){
+            bubbleView.layer.cornerRadius = ALKMessageStyle.receivedBubble.cornerRadius
+            bubbleView.backgroundColor = ALKMessageStyle.receivedBubble.color
+            photoView.layer.cornerRadius = ALKMessageStyle.receivedBubble.cornerRadius
+        }else{
+            photoView.layer.cornerRadius = ALKMessageStyle.receivedBubble.cornerRadius
+            bubbleView.layer.cornerRadius = ALKMessageStyle.receivedBubble.cornerRadius
+        }
     }
     
     override func setupViews() {
@@ -49,9 +56,6 @@ class ALKFriendPhotoCell: ALKPhotoCell {
         avatarImageView.addGestureRecognizer(tapGesture)
 
         contentView.addViewsForAutolayout(views: [avatarImageView,nameLabel])
-        
-        bubbleView.backgroundColor = UIColor.hex8(Color.Background.grayF2.rawValue).withAlphaComponent(0.26)
-        
         nameLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 6).isActive = true
         nameLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 57).isActive = true