+++ Sources/Views/ALKMessageCell.swift
@@ -163,8 +163,14 @@ open class ALKFriendMessageCell: ALKMessageCell {
         super.setupStyle()
 
         nameLabel.setStyle(ALKMessageStyle.displayName)
-        bubbleView.image = bubbleViewImage(for: ALKMessageStyle.receivedBubble.style, isReceiverSide: true,showHangOverImage: false)
-        bubbleView.tintColor = ALKMessageStyle.receivedBubble.color
+        if ALKMessageStyle.receivedBubble.style == .edge {
+            bubbleView.tintColor = UIColor(netHex: 0xF1F0F0)
+            bubbleView.image = bubbleViewImage(for: ALKMessageStyle.receivedBubble.style, isReceiverSide: true,showHangOverImage: false)
+        }else{
+            bubbleView.layer.cornerRadius = ALKMessageStyle.receivedBubble.cornerRadius
+            bubbleView.tintColor = ALKMessageStyle.receivedBubble.color
+            bubbleView.backgroundColor = ALKMessageStyle.receivedBubble.color
+        }
     }
 
     override func update(viewModel: ALKMessageViewModel) {
@@ -223,17 +229,6 @@ open class ALKFriendMessageCell: ALKMessageCell {
         return 6
     }
 
-    // MARK: - ChatMenuCell
-    override func menuWillShow(_ sender: Any) {
-        super.menuWillShow(sender)
-        self.bubbleView.image = bubbleViewImage(for: ALKMessageStyle.receivedBubble.style,isReceiverSide: true,showHangOverImage: true)
-    }
-
-    override func menuWillHide(_ sender: Any) {
-        super.menuWillHide(sender)
-        self.bubbleView.image =  bubbleViewImage(for: ALKMessageStyle.receivedBubble.style,isReceiverSide: true,showHangOverImage: false)
-    }
-
     override class func rowHeigh(viewModel: ALKMessageViewModel,width: CGFloat) -> CGFloat {
 
         // TODO: need to find a better way to calculate the
@@ -253,6 +248,21 @@ open class ALKFriendMessageCell: ALKMessageCell {
 
     @objc private func avatarTappedAction() {
         avatarTapped?()
+    }
+
+    // MARK: - ChatMenuCell
+    override func menuWillShow(_ sender: Any) {
+        super.menuWillShow(sender)
+        if(ALKMessageStyle.receivedBubble.style == .edge){
+            self.bubbleView.image = bubbleViewImage(for: ALKMessageStyle.receivedBubble.style,isReceiverSide: true,showHangOverImage: true)
+        }
+    }
+
+    override func menuWillHide(_ sender: Any) {
+        super.menuWillHide(sender)
+        if(ALKMessageStyle.receivedBubble.style == .edge){
+            self.bubbleView.image =  bubbleViewImage(for: ALKMessageStyle.receivedBubble.style,isReceiverSide: true,showHangOverImage: false)
+        }
     }
 }
 
@@ -382,10 +392,16 @@ open class ALKMyMessageCell: ALKMessageCell {
         timeLabel.bottomAnchor.constraint(equalTo: bubbleView.bottomAnchor, constant: 2).isActive = true
     }
 
-  open  override func setupStyle() {
+    open  override func setupStyle() {
         super.setupStyle()
-        bubbleView.image = bubbleViewImage(for: ALKMessageStyle.sentBubble.style,isReceiverSide: false,showHangOverImage: false)
-        bubbleView.tintColor = ALKMessageStyle.sentBubble.color
+        if ALKMessageStyle.sentBubble.style == .edge {
+            bubbleView.tintColor = UIColor(netHex: 0xF1F0F0)
+            bubbleView.image = bubbleViewImage(for: ALKMessageStyle.sentBubble.style, isReceiverSide: false,showHangOverImage: false)
+        }else{
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.tintColor = ALKMessageStyle.sentBubble.color
+            bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
+        }
     }
 
    open override func update(viewModel: ALKMessageViewModel) {
@@ -420,17 +436,6 @@ open class ALKMyMessageCell: ALKMessageCell {
             stateView.image = UIImage(named: "seen_state_0", in: Bundle.applozic, compatibleWith: nil)
             stateView.tintColor = UIColor.red
         }
-    }
-
-    // MARK: - ChatMenuCell
-    override func menuWillShow(_ sender: Any) {
-        super.menuWillShow(sender)
-        self.bubbleView.image = bubbleViewImage(for: ALKMessageStyle.sentBubble.style,isReceiverSide: false,showHangOverImage: true)
-    }
-
-    override func menuWillHide(_ sender: Any) {
-        super.menuWillHide(sender)
-        self.bubbleView.image =  bubbleViewImage(for: ALKMessageStyle.sentBubble.style,isReceiverSide: false,showHangOverImage: false)
     }
 
     override class func rowHeigh(viewModel: ALKMessageViewModel, width: CGFloat) -> CGFloat {
@@ -475,6 +480,21 @@ open class ALKMyMessageCell: ALKMessageCell {
         previewImageView.constraint(withIdentifier: ConstraintIdentifier.PreviewImage.height)?.constant = 0
         previewImageView.constraint(withIdentifier: ConstraintIdentifier.PreviewImage.width)?.constant = 0
     }
+
+    // MARK: - ChatMenuCell
+    override func menuWillShow(_ sender: Any) {
+        super.menuWillShow(sender)
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            self.bubbleView.image = bubbleViewImage(for: ALKMessageStyle.sentBubble.style,isReceiverSide:false,showHangOverImage: true)
+        }
+    }
+
+    override func menuWillHide(_ sender: Any) {
+        super.menuWillHide(sender)
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            self.bubbleView.image =  bubbleViewImage(for: ALKMessageStyle.sentBubble.style,isReceiverSide: false,showHangOverImage: false)
+        }
+    }
 }
 
 open class ALKMessageCell: ALKChatBaseCell<ALKMessageViewModel>, ALKCopyMenuItemProtocol, ALKReplyMenuItemProtocol {
@@ -493,9 +513,7 @@ open class ALKMessageCell: ALKChatBaseCell<ALKMessageViewModel>, ALKCopyMenuItem
 
     fileprivate var bubbleView: UIImageView = {
         let bv = UIImageView()
-        let image = UIImage.init(named: "chat_bubble_red", in: Bundle.applozic, compatibleWith: nil)
-        bv.tintColor = UIColor(netHex: 0xF1F0F0)
-        bv.image = image?.imageFlippedForRightToLeftLayoutDirection()
+        bv.clipsToBounds = true
         bv.isUserInteractionEnabled = false
         bv.isOpaque = true
         return bv
@@ -622,6 +640,7 @@ open class ALKMessageCell: ALKChatBaseCell<ALKMessageViewModel>, ALKCopyMenuItem
     }
 
 
+
     override class func rowHeigh(viewModel: ALKMessageViewModel,width: CGFloat) -> CGFloat {
 
         var messageHeigh: CGFloat = 0
@@ -690,20 +709,13 @@ open class ALKMessageCell: ALKChatBaseCell<ALKMessageViewModel>, ALKCopyMenuItem
 
     func bubbleViewImage(for style: ALKMessageStyle.BubbleStyle, isReceiverSide: Bool = false,showHangOverImage:Bool) -> UIImage? {
 
-        func getImage(style: ALKMessageStyle.BubbleStyle) -> UIImage? {
-            switch style {
-            case .edge:
-                var imageTitle = showHangOverImage ? "chat_bubble_red_hover":"chat_bubble_red"
+        var imageTitle = showHangOverImage ? "chat_bubble_red_hover":"chat_bubble_red"
+        // We can rotate the above image but loading the required
+        // image would be faster and we already have both the images.
+        if isReceiverSide {imageTitle = showHangOverImage ? "chat_bubble_grey_hover":"chat_bubble_grey"}
 
-                // We can rotate the above image but loading the required
-                // image would be faster and we already have both the images.
-                if isReceiverSide {imageTitle = showHangOverImage ? "chat_bubble_grey_hover":"chat_bubble_grey"}
-                return UIImage.init(named: imageTitle, in: Bundle.applozic, compatibleWith: nil)
-            case .round:
-                return UIImage.init(named: "chat_bubble_rounded", in: Bundle.applozic, compatibleWith: nil)
-            }
-        }
-        guard let bubbleImage = getImage(style: style) else {return nil}
+        guard let bubbleImage = UIImage.init(named: imageTitle, in: Bundle.applozic, compatibleWith: nil)
+            else {return nil}
 
         // This API is from the Kingfisher so instead of directly using
         // imageFlippedForRightToLeftLayoutDirection() we are using this as it handles