+++ Sources/Views/ALKMyGenericListCell.swift
@@ -41,9 +41,7 @@ class ALKMyGenericListCell: ALKChatBaseCell<ALKMessageViewModel> {
     
     fileprivate var bubbleView: UIImageView = {
         let bv = UIImageView()
-        let image = UIImage.init(named: "chat_bubble_rounded", in: Bundle.applozic, compatibleWith: nil)
-        bv.tintColor = UIColor(netHex: 0xF1F0F0)
-        bv.image = image?.imageFlippedForRightToLeftLayoutDirection()
+        bv.clipsToBounds = true
         bv.isUserInteractionEnabled = false
         bv.isOpaque = true
         return bv
@@ -154,6 +152,19 @@ class ALKMyGenericListCell: ALKChatBaseCell<ALKMessageViewModel> {
         setupConstraints()
         backgroundColor = .clear
     }
+
+    override func setupStyle() {
+        super.setupStyle()
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            let image = UIImage.init(named: "chat_bubble_rounded", in: Bundle.applozic, compatibleWith: nil)
+            bubbleView.tintColor = UIColor(netHex: 0xF1F0F0)
+            bubbleView.image = image?.imageFlippedForRightToLeftLayoutDirection()
+        }else{
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.tintColor = ALKMessageStyle.sentBubble.color
+            bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
+        }
+     }
     
     private func setupConstraints() {
         let view = contentView