+++ Sources/Views/ALKMyLocationCell.swift
@@ -22,9 +22,6 @@ final class ALKMyLocationCell: ALKLocationCell {
     // MARK: - Lifecycle
     override func setupViews() {
         super.setupViews()
-
-        bubbleView.backgroundColor = UIColor.background(.redC0)
-
         // add view to contenview and setup constraint
         contentView.addViewsForAutolayout(views: [stateView])
 
@@ -56,6 +53,18 @@ final class ALKMyLocationCell: ALKLocationCell {
         } else {
             stateView.image = UIImage(named: "seen_state_0", in: Bundle.applozic, compatibleWith: nil)
             stateView.tintColor = UIColor.red
+        }
+    }
+
+    override func setupStyle() {
+        super.setupStyle()
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            bubbleView.backgroundColor = UIColor.background(.redC0)
+            bubbleView.layer.cornerRadius = 12
+        }else{
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.tintColor = ALKMessageStyle.sentBubble.color
+            bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
         }
     }