+++ Sources/Views/ALKMyMessageView.swift
@@ -50,9 +50,7 @@ class ALKMyMessageView: UIView {
 
     fileprivate var bubbleView: UIImageView = {
         let bv = UIImageView()
-        let image = UIImage.init(named: "chat_bubble_rounded", in: Bundle.applozic, compatibleWith: nil)
-        bv.tintColor = UIColor(netHex: 0xF1F0F0)
-        bv.image = image?.imageFlippedForRightToLeftLayoutDirection()
+        bv.clipsToBounds = true
         bv.isUserInteractionEnabled = false
         bv.isOpaque = true
         return bv
@@ -61,10 +59,23 @@ class ALKMyMessageView: UIView {
     init() {
         super.init(frame: .zero)
         setupConstraints()
+        setupStyle()
     }
 
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
+    }
+
+    func setupStyle() {
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            let image = UIImage.init(named: "chat_bubble_rounded", in: Bundle.applozic, compatibleWith: nil)
+            bubbleView.tintColor = UIColor(netHex: 0xF1F0F0)
+            bubbleView.image = image?.imageFlippedForRightToLeftLayoutDirection()
+        }else{
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.tintColor = ALKMessageStyle.sentBubble.color
+            bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
+        }
     }
 
     func update(viewModel: ALKMessageViewModel) {