+++ Sources/Views/ALKMyPhotoLandscapeCell.swift
@@ -32,8 +32,6 @@ final class ALKMyPhotoLandscapeCell: ALKPhotoCell {
         photoView.widthAnchor.constraint(equalToConstant: width*0.64).isActive = true
         photoView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -6).isActive = true
         
-        bubbleView.backgroundColor = UIColor.hex8(Color.Background.grayF2.rawValue).withAlphaComponent(0.26)
-        
         fileSizeLabel.rightAnchor.constraint(equalTo: bubbleView.rightAnchor, constant: 0).isActive = true
         
         stateView.widthAnchor.constraint(equalToConstant: 17.0).isActive = true
@@ -61,5 +59,17 @@ final class ALKMyPhotoLandscapeCell: ALKPhotoCell {
     
     override class func bottomPadding() -> CGFloat {
         return 6
+    }
+
+    override func setupStyle() {
+        super.setupStyle()
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
+            photoView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+        }else{
+            photoView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+        }
     }
 }