+++ Sources/Views/ALKMyPhotoPortalCell.swift
@@ -29,8 +29,6 @@ final class ALKMyPhotoPortalCell: ALKPhotoCell {
         super.setupViews()
         
         let width = UIScreen.main.bounds.width
-
-        
         contentView.addViewsForAutolayout(views: [stateView])
         
         photoView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 6).isActive = true
@@ -40,8 +38,6 @@ final class ALKMyPhotoPortalCell: ALKPhotoCell {
         
         photoView.widthAnchor.constraint(equalToConstant: width*0.48).isActive = true
         photoView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -6).isActive = true
-        
-        bubbleView.backgroundColor = UIColor.hex8(Color.Background.grayF2.rawValue).withAlphaComponent(0.26)
         
         fileSizeLabel.rightAnchor.constraint(equalTo: bubbleView.rightAnchor, constant: 0).isActive = true
         
@@ -77,4 +73,15 @@ final class ALKMyPhotoPortalCell: ALKPhotoCell {
         return 6
     }
 
+    override func setupStyle() {
+        super.setupStyle()
+        if(ALKMessageStyle.sentBubble.style == .edge){
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
+            photoView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+        }else{
+            photoView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+            bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius
+        }
+    }
 }