+++ Sources/Views/ALKPhotoCell.swift
@@ -20,7 +20,6 @@ class ALKPhotoCell: ALKChatBaseCell<ALKMessageViewModel>,
         mv.backgroundColor = .clear
         mv.contentMode = .scaleAspectFill
         mv.clipsToBounds = true
-        mv.layer.cornerRadius = 12
         return mv
     }()
 
@@ -36,8 +35,6 @@ class ALKPhotoCell: ALKChatBaseCell<ALKMessageViewModel>,
 
     var bubbleView: UIView = {
         let bv = UIView()
-        bv.backgroundColor = .gray
-        bv.layer.cornerRadius = 12
         bv.isUserInteractionEnabled = false
         return bv
     }()