+++ Sources/Views/ALKVideoCell.swift
@@ -25,7 +25,6 @@ class ALKVideoCell: ALKChatBaseCell<ALKMessageViewModel>,
         mv.backgroundColor = .clear
         mv.contentMode = .scaleAspectFill
         mv.clipsToBounds = true
-        mv.layer.cornerRadius = 12
         return mv
     }()
 
@@ -71,8 +70,7 @@ class ALKVideoCell: ALKChatBaseCell<ALKMessageViewModel>,
 
     var bubbleView: UIView = {
         let bv = UIView()
-        bv.backgroundColor = .gray
-        bv.layer.cornerRadius = 12
+        bv.clipsToBounds = true;
         bv.isUserInteractionEnabled = false
         return bv
     }()